Lidar Packet
In Main, lidar data is continuously received unless the user stops it.

With the LidarReader() class, data is received accurately through serial communication.

The read bytes are written to the .txt file from the moment it finds the header.

In the LidarPacket() class, packet variables are kept appropriately.

step is calculated.